modifier credit et some joint modifier montant et montant par mois with float

+++ src/main/java/tn/esprit/spring/entity/Credit.java
@@ -34,11 +34,11 @@ public class Credit implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long creditId;
-	private Long montant;
+	private float montant;
 	private String iban;
 	private Long restapayer;
 	private int nbrdumois;
-	private Long montantparmois;
+	private float montantparmois;
 	@Temporal(TemporalType.DATE)
 	private Date dateDebut;
 	@Temporal(TemporalType.DATE)
